clinidoc/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
pydantic
pytest


